Move document modules so they load in debugger
- Where they were previously, if you went into a debugger and tried to call anything that called a SolrDocument, you would get the error `NameError Exception: uninitialized constant SolrDocument::Alma`. Now you can run in debugger successfully.
