+++ app/models/concerns/orangelight/document/standard_numbers.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+module Orangelight
+  module Document
+    # Standard Numbers functionality for SolrDocument
+    module StandardNumbers
+      def standard_numbers?
+        standard_number_fields.any? { |field| key? field }
+      end
+
+      private
+
+        def standard_number_fields
+          %w[lccn_s isbn_s issn_s oclc_s]
+        end
+    end
+  end
+end
