+++ app/models/solr_document.rb
@@ -3,8 +3,8 @@
 class SolrDocument
   include Blacklight::Solr::Document
   include Orangelight::Document::Export
-  include SolrDocument::Alma
-  include SolrDocument::StandardNumbers
+  include Orangelight::Document::Alma
+  include Orangelight::Document::StandardNumbers
 
   # Explicitly required for sneakers
   include Blacklight::Document::Extensions
